// src/api/models/user-registration.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UserRegistration
 */
export interface UserRegistration {
    /**
     * 
     * @type {string}
     * @memberof UserRegistration
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof UserRegistration
     */
    'first_name': string;
    /**
     * 
     * @type {string}
     * @memberof UserRegistration
     */
    'last_name': string;
    /**
     * 
     * @type {string}
     * @memberof UserRegistration
     */
    'password': string;
    /**
     * 
     * @type {string}
     * @memberof UserRegistration
     */
    'password2': string;
}